Use (default) namespace from context
